% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CliMaps-package.r, R/CliMapsLink.R
\docType{package}
\name{CliMaps}
\alias{CliMaps}
\alias{CliMaps-package}
\title{CliMaps.}
\usage{
CliMaps(sp, lon.grid, lat.grid, data, names, DT = FALSE, IDvar,
  grid.tolerance = 1e-04)
}
\arguments{
\item{sp}{The spatial data. The data can be of class spatialpolygon, spatialpolygondataframe, or a data.frame with with columns for longitude and latitude. Each spatial unit in sp will be linked to the gridded data.}

\item{lon.grid}{A numeric vector of numeric longitudes for the centroids of each gridcell}

\item{lat.grid}{A numeric vector of numeric lattitudes for the centroids of each gridcell}

\item{data}{Option array of of data. The dimensions should be 1) longitude, 2) latitude, and 2) time. If data is provided than CliMaps both links the data and creates the time series. If data is not provided then CliMaps only creates the link.}

\item{names}{An option list of names for the dimensions of data. The names for time will be the colnames of the timeseries. The vectors in the list should be named lon, lat, and time.}

\item{DT}{Ligical indicating if a data.table should be returned. If FALSE (default) a data.frame will be returned.}

\item{IDvar}{A optional variable to be used as an ID if the spatial data is provided as a data.frame. This is not used if an opject of class sp is provided.}

\item{grid.tolerance}{This is a tolerance level used to make sure the grid is evenly spaced. If the grid is not evenly spaced it can still be treated as evenly spaced by setting a high tolderance.}
}
\description{
CliMaps.

This function links each geographical unites to grid cells. The linking is area weighted to each grid.
}
\author{
Ander Wilson
}

