% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CliMapsTS.R
\name{CliMapsTS}
\alias{CliMapsTS}
\title{CliMapsTS}
\usage{
CliMapsTS(link, data, names)
}
\arguments{
\item{link}{A link from from the function CliMaps.}

\item{data}{An array of of data. The dimensions should be 1) longitude, 2) latitude, and 2) time. If data is provided than CliMaps both links the data and creates the time series. If data is not provided then CliMaps only creates the link.}

\item{names}{A list of names for the dimensions of data. The names for time will be the colnames of the timeseries. The vectors in the list should be named lon, lat, and time.}
}
\description{
This function takes a link from the function CliMaps and an array of gridded data and creates a data.frame of time series data.
}
\author{
Ander Wilson
}

