% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.tiger.R
\name{get.tiger}
\alias{get.tiger}
\title{get.tiger}
\usage{
get.tiger(fn, extent, layer, yy, src = "tiger", version = 2010,
  loc.path = "")
}
\arguments{
\item{fn}{The file name to be downloaded. This takes the format of tl_<version>_<extent>_<layer><year>. Either fn or the version, year and extent must be provided.}

\item{extent}{A parent geography id.  This can be a five-digit county fips, two-digit state fips, or "us" for the whole county. It should be provided as a character. Not all extents are available for all layers.}

\item{layer}{The layer of the file to be downloaded. Examples include: state, county, tract, TABBLOCK (census block), or ZCTA5 (five digit sip code tabulation area).}

\item{yy}{The two-digit year provided as a character.  This is is most likely "00" or "10" for the two decennial census years.}

\item{src}{The source of the file. Only TIGER is available now. This will be extended to the smaller GENZ files in the future.}

\item{version}{The version that hsould be used. The default is 2010. This function has limited capabilities for versions before 2010.}

\item{loc.path}{The local path where the data should be downloaded to.}
}
\description{
This function downloads TIGER shapefiles from the US Census. It is designed for version 2010 and later. In most cases it will not work for earlier versions. This only needs to be done once.
}
\examples{
# Download the 2010 state shape file for the entire US.
# This will download an approximately 8.3Mb file and may take a few minutes.
# This only needs to be done once.
tiger.info <- get.tiger(fn = "tl_2010_us_state10")
}
\author{
Ander Wilson
}

