% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.tiger.R
\name{load.tiger}
\alias{load.tiger}
\title{load.tiger}
\usage{
load.tiger(path, fn)
}
\arguments{
\item{path}{A local path where the TIGER or GENZ shape data are stored.}

\item{fn}{A file name of the form tl_<version>_<extent>_<layer><year>.}
}
\description{
This function takes a link from the function CliMaps and an array of gridded data and creates a data.frame of time series data.
}
\examples{
# Download the 2010 state shape file for the entire US.
# This will download an approximately 8.3Mb file and may take a few minutes.
# This only needs to be done once.
tiger.info <- get.tiger(fn = "tl_2010_us_state10")

# load the shape file
sp <- load.tiger(tiger.info$path,tiger.info$fn)
}
\author{
Ander Wilson
}

